# resiresi-backend/openapi.yaml
openapi: "3.1.0"
info:
  title: ResiResi Restaurant Reservation API
  version: "1.0.0"
  description: |
    Backend reservation API for the ResiResi platform. Used as the APIblaze proxy reference origin.
    Authenticate with either `Authorization: Bearer <key>` or `x-api-key: <key>`.
servers:
  - url: https://backend.resiresi.com/v1
    description: Production
  - url: http://localhost:8080/v1
    description: Local development
security:
  - ApiKeyAuth: []
  - BearerAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
              example: not_found
            message:
              type: string
              example: Restaurant not found.
            request_id:
              type: string
              format: uuid
            details:
              type: object
              description: Zod flatten() output on validation errors
    OpenHour:
      type: object
      required: [day_of_week, open, close]
      properties:
        day_of_week:
          type: integer
          minimum: 0
          maximum: 6
          description: "0=Sunday … 6=Saturday"
        open:
          type: string
          pattern: '^\d{2}:\d{2}$'
          example: "11:00"
        close:
          type: string
          pattern: '^\d{2}:\d{2}$'
          example: "22:00"
    Restaurant:
      type: object
      required: [id, slug, name, timezone, open_hours, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
          example: le-bernardin
        name:
          type: string
          example: Le Bernardin
        timezone:
          type: string
          example: America/New_York
        open_hours:
          type: array
          items:
            $ref: '#/components/schemas/OpenHour'
        address:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Table:
      type: object
      required: [id, restaurant_id, label, capacity, active]
      properties:
        id:
          type: string
          format: uuid
        restaurant_id:
          type: string
          format: uuid
        label:
          type: string
          example: B1
        capacity:
          type: integer
          minimum: 1
        active:
          type: boolean
...
    RestaurantList:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Restaurant'
    TableList:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Table'
...
paths:
...
  /restaurants:
    get:
      operationId: listRestaurants
      summary: List restaurants
      tags: [Restaurants]
      responses:
        "200":
          description: List of restaurants
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestaurantList'
        "401":
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      operationId: createRestaurant
      summary: Create restaurant (platform only)
      tags: [Restaurants]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slug, name]
              properties:
                slug:
                  type: string
                  pattern: '^[a-z0-9-]+$'
                name:
                  type: string
                timezone:
                  type: string
                  default: America/Detroit
                open_hours:
                  type: array
                  items:
                    $ref: '#/components/schemas/OpenHour'
                address:
                  type: string
                phone:
                  type: string
      responses:
        "201":
          description: Restaurant created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restaurant'
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: Forbidden — requires platform key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "409":
          description: Slug already taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /restaurants/{restaurantId}:
    parameters:
      - name: restaurantId
        in: path
        required: true
        schema:
          type: string

    get:
      operationId: getRestaurant
      summary: Get restaurant by ID or slug
      tags: [Restaurants]
      responses:
        "200":
          description: Restaurant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restaurant'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      operationId: updateRestaurant
      summary: Update restaurant (owner/manager)
      tags: [Restaurants]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                timezone:
                  type: string
                open_hours:
                  type: array
                  items:
                    $ref: '#/components/schemas/OpenHour'
                address:
                  type: string
                phone:
                  type: string
      responses:
        "200":
          description: Updated restaurant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restaurant'
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /restaurants/{restaurantId}/tables:
    parameters:
      - name: restaurantId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      operationId: listTables
      summary: List tables
      tags: [Tables]
      responses:
        "200":
          description: Tables
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableList'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      operationId: createTable
      summary: Create table (owner/manager)
      tags: [Tables]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [label, capacity]
              properties:
                label:
                  type: string
                capacity:
                  type: integer
                  minimum: 1
                active:
                  type: boolean
                  default: true
      responses:
        "201":
          description: Table created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Table'
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "409":
          description: Table label already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /restaurants/{restaurantId}/tables/{tableId}:
    parameters:
      - name: restaurantId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: tableId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    patch:
      operationId: updateTable
      summary: Update table (owner/manager)
      tags: [Tables]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                label:
                  type: string
                capacity:
                  type: integer
                  minimum: 1
                active:
                  type: boolean
      responses:
        "200":
          description: Updated table
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Table'
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "409":
          description: Table label conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      operationId: deleteTable
      summary: Delete table (owner/manager)
      tags: [Tables]
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...